Fix: string filter was broken
